Breakout frame title bar into components
